# docs/openapi/admin-v1.yaml
openapi: 3.1.0
info:
  title: BlakTail Admin API
  version: "1"
  description: |
    Organisation-scoped automation API. Distinct from node protocol routes
    and console HMAC assertions. Compatibility is additive inside `/api/v1`.
servers:
  - url: https://coord.example
security:
  - bearerAuth: []
  - organisationHeader: []
paths:
...
  /api/v1/devices:
    get:
      summary: List devices
      parameters:
        - $ref: "#/components/parameters/Organisation"
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string, enum: [active, online, offline, revoked, deleted, expired, ephemeral] }
        - in: query
          name: include_deleted
          schema: { type: boolean }
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 200 }
        - in: query
          name: before
          schema: { type: string }
      responses:
        "200":
          description: Envelope of devices
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: bta_ automation token. Node tokens and join keys are rejected.
    organisationHeader:
      type: apiKey
      in: header
      name: X-BlakTail-Organisation
  parameters:
    Organisation:
      in: header
      name: X-BlakTail-Organisation
      required: true
      schema: { type: string, format: uuid }
...